feat: added state setting logic from the schedule courses
