+++ client/src/lib/utils.ts
@@ -1,0 +1,6 @@
+/* eslint-disable-next-line @typescript-eslint/no-explicit-any */
+const classNames = (...classes: any[]) => {
+  return classes.filter(Boolean).join(" ");
+};
+
+export { classNames };
